Update src/components file structure. Add missing unit & integration tests.
