+++ src/components/index.ts
@@ -1,15 +1,15 @@
-export { AuroraBackground } from './AuroraBackground/AuroraBackground';
-export { Card } from './Card/Card';
-export { ContactForm } from './ContactForm/ContactForm';
-export { ExperienceTimeline } from './ExperienceTimeline/ExperienceTimeline';
-export { Header } from './Header/Header';
-export { Hero } from './Hero/Hero';
-export { HexGrid } from './HexGrid/HexGrid';
-export { Highlights } from './Highlights/Highlights';
-export { Footer } from './Footer/Footer';
-export { Logo } from './Logo/Logo';
-export { RevealSection } from './RevealSection/RevealSection';
-export { SectionHeading } from './SectionHeading/SectionHeading';
-export { SkillsGrid } from './SkillsGrid/SkillsGrid';
-export { SocialLinks } from './SocialLinks/SocialLinks';
-export type { Project } from './Card/Card';
+export { AuroraBackground } from './ui/AuroraBackground/AuroraBackground';
+export { Card } from './sections/Card/Card';
+export { ContactForm } from './sections/ContactForm/ContactForm';
+export { ExperienceTimeline } from './sections/ExperienceTimeline/ExperienceTimeline';
+export { Header } from './layout/Header/Header';
+export { Hero } from './sections/Hero/Hero';
+export { HexGrid } from './ui/HexGrid/HexGrid';
+export { Highlights } from './sections/Highlights/Highlights';
+export { Footer } from './layout/Footer/Footer';
+export { Logo } from './ui/Logo/Logo';
+export { RevealSection } from './ui/RevealSection/RevealSection';
+export { SectionHeading } from './ui/SectionHeading/SectionHeading';
+export { SkillsGrid } from './sections/SkillsGrid/SkillsGrid';
+export { SocialLinks } from './ui/SocialLinks/SocialLinks';
+export type { Project } from './sections/Card/Card';
