+++ src/components/ui/AuroraBackground/AuroraBackground.scss
